changed
Changed the worst value to two. Could change the two class case to calculate it as in the more than two class case, but probably this will lead to too much confusion for the user, so I leave it like that. 
That brier is binary can be seen in the tutorial page, but renaming it, would make it clearer.