create a thesthtat for h2oautoencoder, analogous to h2ogbm. BUT:
In test_that - “oneclass_h2oautoencoder”, reproducible = TRUE and seed= 1234 need to be set in parset.list before the for-block, instead of setting them in the for-block like in h2ogmb, otherwise testProbParsets return error because of different prediction values.

In test_that-(“class names are integers and probabilities predicted (#1787)"
activation function needs to be set to “Tanh” instead of the default “Rectifire”, otherwise resample() return an error:
Error: DistributedException from localhost/127.0.0.1:54321, caused by java.lang.UnsupportedOperationException:
(This error occurs very often when using h2o, there are no clear and not a lot explanation in the web how to solve it or why it happens, but it seems like to be related with the case if the model is “unstable”. The activation function “tanh” has a natural bound and suitable to “control” the instability.